+++ playground/.gitignore
@@ -1,7 +1,7 @@
 # See https://help.github.com/articles/ignoring-files/ for more about ignoring files.
 
-evaluation_data/*
-!evaluation_data/.gitkeep
+data/evaluation/*
+!data/evaluation/.gitkeep
 
 # dependencies
 /node_modules
