+++ playground/data/example/exercise-1.json
@@ -8,15 +8,15 @@
     "problem_statement": "-> file:problem-statement.md",
     "grading_instructions": "-> file:grading-instructions.md",
     "programming_language": "java",
-    "solution_repository_url": "{{exampleExercisesUrl}}/1/solution.zip",
-    "template_repository_url": "{{exampleExercisesUrl}}/1/template.zip",
-    "tests_repository_url": "{{exampleExercisesUrl}}/1/tests.zip",
+    "solution_repository_url": "{{exampleDataUrl}}/1/solution.zip",
+    "template_repository_url": "{{exampleDataUrl}}/1/template.zip",
+    "tests_repository_url": "{{exampleDataUrl}}/1/tests.zip",
     "meta": {},
 
     "submissions": [
         {
             "id": 101,
-            "repository_url": "{{exampleExercisesUrl}}/1/submissions/101.zip",
+            "repository_url": "{{exampleDataUrl}}/1/submissions/101.zip",
             "student_id": 1,
             "meta": {},
 
@@ -39,13 +39,13 @@
         },
         {
             "id": 102,
-            "repository_url": "{{exampleExercisesUrl}}/1/submissions/102.zip",
+            "repository_url": "{{exampleDataUrl}}/1/submissions/102.zip",
             "student_id": 2,
             "meta": {}
         },
         {
             "id": 103,
-            "repository_url": "{{exampleExercisesUrl}}/1/submissions/103.zip",
+            "repository_url": "{{exampleDataUrl}}/1/submissions/103.zip",
             "student_id": 3,
             "meta": {}
         }
