+++ playground/data/example/exercise-1/solution/gradle/wrapper/gradle-wrapper.properties
