+++ playground/data/example/exercise-1/solution/settings.gradle
