+++ playground/data/example/exercise-1/solution/src/mttpe1/Policy.java
