+++ playground/data/example/exercise-1/submissions/101/BinarySearch.java
