+++ playground/data/example/exercise-1/submissions/103/gradle/wrapper/gradle-wrapper.properties
