+++ playground/data/example/exercise-1/submissions/103/settings.gradle
