+++ playground/data/example/exercise-1/submissions/103/src/mttpe1/SortStrategy.java
