+++ playground/data/example/exercise-1/template/build.gradle
