+++ playground/data/example/exercise-1/template/gradle/wrapper/gradle-wrapper.properties
