+++ playground/data/example/exercise-1/template/gradlew.bat
