+++ playground/data/example/exercise-1/template/settings.gradle
