+++ playground/data/example/exercise-1/template/src/mttpe1/MergeSort.java
