+++ playground/data/example/exercise-1/tests/gradle/wrapper/gradle-wrapper.properties
