+++ playground/data/example/exercise-1/tests/settings.gradle
