+++ playground/src/pages/api/exercises.ts
@@ -1,6 +1,6 @@
 import type {NextApiRequest, NextApiResponse} from 'next';
 import type { Exercise } from '@/model/exercise';
-import {getExampleExercises} from "@/helpers/get_examples";
+import {getExampleExercises} from "@/helpers/get_data";
 import getOriginFromRequest from '@/helpers/origin_from_req';
 
 export default function handler(
